requirements.txt
bcrypt==2.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.0
Django==1.8.4
django-allauth==0.23.0
django-appconf==1.0.1
django-autocomplete-light==2.2.5
django-autoslug==1.8.0
django-compressor==1.5
django-easy-pjax==1.2.0
django-extensions==1.5.6
django-filter==0.11.0
django-flat-theme==1.1.1
django-grappelli==2.7.1
django-htmlmin==0.7.0
django-polymorphic==0.7.1
django-rest-auth==0.5.0
django-suit==0.3a1
django-test-pep8==0.1
djangorestframework==3.2.3
html5lib==0.999999
jsonfield==1.0.3
mysqlclient==1.3.6
oauthlib==1.0.3
pep8==1.6.2
Pillow==2.9.0
pycparser==2.14
python-openid==2.2.5
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
versiontools==1.9.1
wheel==0.24.0
